+++ server/tests/backends/pypika_translator/test_redshift_translator.py
@@ -105,7 +105,10 @@ def get_top_query(sort_order, previous_step, group, rank_on, selected_columns, l
         .orderby(age_field, order=Order.asc if sort_order == "asc" else Order.desc)
     )
     expected_query = (
-        Query.from_(rank_select).where(Field("row_number") == limit).select(*selected_columns)
+        Query.from_(rank_select)
+        .where(Field("row_number") == limit)
+        .orderby(*(Field(f) for f in [group, "row_number"]), order=Order.asc)
+        .select(*selected_columns)
     )
     return expected_query