+++ server/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Fix: Ensure RowNumber() has an alias in the `top` step of the PyPika translator
+- Fix: Ensure the `top` step of the PyPika translator returns consistent results
 
 ## [0.16.1] - 2022-06-22